refactor: llm_report_service로 이름 변경
